@@ -1,253 +1,204 @@
 {
   "name": "libp2p",
-  "version": "0.35.6",
+  "version": "0.36.2",
   "description": "JavaScript implementation of libp2p, a modular peer to peer network stack",
-  "leadMaintainer": "Jacob Heun <jacobheun@gmail.com>",
-  "main": "src/index.js",
-  "types": "dist/src/index.d.ts",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/libp2p/js-libp2p.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p/issues"
+  },
+  "keywords": [
+    "IPFS",
+    "libp2p",
+    "network",
+    "p2p",
+    "peer",
+    "peer-to-peer"
+  ],
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
   "typesVersions": {
     "*": {
+      "*": [
+        "*",
+        "dist/*",
+        "dist/src/*",
+        "dist/src/*/index"
+      ],
       "src/*": [
+        "*",
+        "dist/*",
         "dist/src/*",
         "dist/src/*/index"
       ]
     }
   },
   "files": [
-    "dist",
-    "src"
+    "src",
+    "dist/src",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
   ],
-  "scripts": {
-    "lint": "aegir lint",
-    "build": "aegir build",
-    "build:proto": "npm run build:proto:circuit && npm run build:proto:identify && npm run build:proto:plaintext && npm run build:proto:address-book && npm run build:proto:proto-book && npm run build:proto:peer-record && npm run build:proto:envelope",
-    "build:proto:circuit": "pbjs -t static-module -w commonjs -r libp2p-circuit --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/circuit/protocol/index.js ./src/circuit/protocol/index.proto",
-    "build:proto:identify": "pbjs -t static-module -w commonjs -r libp2p-identify --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/identify/message.js ./src/identify/message.proto",
-    "build:proto:plaintext": "pbjs -t static-module -w commonjs -r libp2p-plaintext --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/insecure/proto.js ./src/insecure/proto.proto",
-    "build:proto:address-book": "pbjs -t static-module -w commonjs -r libp2p-address-book --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/peer-store/persistent/pb/address-book.js ./src/peer-store/persistent/pb/address-book.proto",
-    "build:proto:proto-book": "pbjs -t static-module -w commonjs -r libp2p-proto-book --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/peer-store/persistent/pb/proto-book.js ./src/peer-store/persistent/pb/proto-book.proto",
-    "build:proto:peer-record": "pbjs -t static-module -w commonjs -r libp2p-peer-record --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/record/peer-record/peer-record.js ./src/record/peer-record/peer-record.proto",
-    "build:proto:envelope": "pbjs -t static-module -w commonjs -r libp2p-envelope --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/record/envelope/envelope.js ./src/record/envelope/envelope.proto",
-    "build:proto-types": "npm run build:proto-types:circuit && npm run build:proto-types:identify && npm run build:proto-types:plaintext && npm run build:proto-types:address-book && npm run build:proto-types:proto-book && npm run build:proto-types:peer-record && npm run build:proto-types:envelope",
-    "build:proto-types:circuit": "pbts -o src/circuit/protocol/index.d.ts src/circuit/protocol/index.js",
-    "build:proto-types:identify": "pbts -o src/identify/message.d.ts src/identify/message.js",
-    "build:proto-types:plaintext": "pbts -o src/insecure/proto.d.ts src/insecure/proto.js",
-    "build:proto-types:address-book": "pbts -o src/peer-store/persistent/pb/address-book.d.ts src/peer-store/persistent/pb/address-book.js",
-    "build:proto-types:proto-book": "pbts -o src/peer-store/persistent/pb/proto-book.d.ts src/peer-store/persistent/pb/proto-book.js",
-    "build:proto-types:peer-record": "pbts -o src/record/peer-record/peer-record.d.ts src/record/peer-record/peer-record.js",
-    "build:proto-types:envelope": "pbts -o src/record/envelope/envelope.d.ts src/record/envelope/envelope.js",
-    "test": "aegir test",
-    "test:ts": "aegir build --no-bundle && npm run test --prefix test/ts-use",
-    "test:node": "aegir test -t node -f \"./test/**/*.{node,spec}.js\"",
-    "test:browser": "aegir test -t browser",
-    "test:examples": "cd examples && npm run test:all",
-    "test:interop": "LIBP2P_JS=$PWD npx aegir test -t node -f ./node_modules/libp2p-interop/test/*",
-    "prepare": "aegir build --no-bundle",
-    "release": "aegir release -t node -t browser",
-    "release-minor": "aegir release --type minor -t node -t browser",
-    "release-major": "aegir release --type major -t node -t browser",
-    "coverage": "nyc --reporter=text --reporter=lcov npm run test:node"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/libp2p/js-libp2p.git"
-  },
-  "keywords": [
-    "libp2p",
-    "network",
-    "p2p",
-    "peer",
-    "peer-to-peer",
-    "IPFS"
-  ],
-  "bugs": {
-    "url": "https://github.com/libp2p/js-libp2p/issues"
-  },
-  "homepage": "https://libp2p.io",
-  "license": "MIT",
-  "engines": {
-    "node": ">=15.0.0"
-  },
-  "browser": {
-    "nat-api": false
+  "exports": {
+    ".": {
+      "import": "./dist/src/index.js"
+    },
+    "./insecure": {
+      "import": "./dist/src/insecure/index.js"
+    },
+    "./pnet": {
+      "import": "./dist/src/pnet/index.js"
+    },
+    "./pnet/generate": {
+      "import": "./dist/src/pnet/key-generator.js"
+    },
+    "./transport-manager": {
+      "import": "./dist/src/transport-manager.js"
+    }
   },
   "eslintConfig": {
     "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    },
     "ignorePatterns": [
       "!.aegir.js",
       "test/ts-use",
       "*.d.ts"
     ]
   },
+  "scripts": {
+    "lint": "aegir lint",
+    "build": "tsc",
+    "postbuild": "mkdirp dist/src/circuit/pb dist/src/fetch/pb dist/src/identify/pb dist/src/insecure/pb && cp src/circuit/pb/*.js src/circuit/pb/*.d.ts dist/src/circuit/pb && cp src/fetch/pb/*.js src/fetch/pb/*.d.ts dist/src/fetch/pb && cp src/identify/pb/*.js src/identify/pb/*.d.ts dist/src/identify/pb && cp src/insecure/pb/*.js src/insecure/pb/*.d.ts dist/src/insecure/pb",
+    "generate": "run-s generate:proto:* generate:proto-types:*",
+    "generate:proto:circuit": "pbjs -t static-module -w es6 -r libp2p-circuit --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/circuit/protocol/index.js ./src/circuit/protocol/index.proto",
+    "generate:proto:fetch": "pbjs -t static-module -w es6 -r libp2p-fetch --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/fetch/proto.js ./src/fetch/proto.proto",
+    "generate:proto:identify": "pbjs -t static-module -w es6 -r libp2p-identify --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/identify/message.js ./src/identify/message.proto",
+    "generate:proto:plaintext": "pbjs -t static-module -w es6 -r libp2p-plaintext --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/insecure/proto.js ./src/insecure/proto.proto",
+    "generate:proto-types:circuit": "pbts -o src/circuit/protocol/index.d.ts src/circuit/protocol/index.js",
+    "generate:proto-types:fetch": "pbts -o src/fetch/proto.d.ts src/fetch/proto.js",
+    "generate:proto-types:identify": "pbts -o src/identify/message.d.ts src/identify/message.js",
+    "generate:proto-types:plaintext": "pbts -o src/insecure/proto.d.ts src/insecure/proto.js",
+    "pretest": "npm run build",
+    "test": "aegir test",
+    "test:node": "npm run test -- -t node -f \"./dist/test/**/*.{node,spec}.js\" --cov",
+    "test:chrome": "npm run test -- -t browser -f \"./dist/test/**/*.spec.js\" --cov",
+    "test:chrome-webworker": "npm run test -- -t webworker -f \"./dist/test/**/*.spec.js\"",
+    "test:firefox": "npm run test -- -t browser -f \"./dist/test/**/*.spec.js\" -- --browser firefox",
+    "test:firefox-webworker": "npm run test -- -t webworker -f \"./dist/test/**/*.spec.js\" -- --browser firefox",
+    "test:examples": "cd examples && npm run test:all",
+    "test:interop": "npm run test -- -t node -f dist/test/interop.js"
+  },
   "dependencies": {
-    "@vascosantos/moving-average": "^1.1.0",
-    "abort-controller": "^3.0.0",
-    "abortable-iterator": "^3.0.0",
-    "aggregate-error": "^3.1.0",
-    "any-signal": "^2.1.1",
+    "@achingbrain/nat-port-mapper": "^1.0.0",
+    "@libp2p/connection": "^1.1.4",
+    "@libp2p/crypto": "^0.22.9",
+    "@libp2p/interfaces": "^1.3.17",
+    "@libp2p/multistream-select": "^1.0.3",
+    "@libp2p/peer-id": "^1.1.8",
+    "@libp2p/peer-id-factory": "^1.0.8",
+    "@libp2p/peer-store": "^1.0.6",
+    "@libp2p/utils": "^1.0.9",
+    "@multiformats/mafmt": "^11.0.2",
+    "@multiformats/multiaddr": "^10.1.8",
+    "abortable-iterator": "^4.0.2",
+    "aggregate-error": "^4.0.0",
+    "any-signal": "^3.0.0",
     "bignumber.js": "^9.0.1",
     "class-is": "^1.1.0",
-    "debug": "^4.3.1",
-    "err-code": "^3.0.0",
-    "es6-promisify": "^7.0.0",
+    "datastore-core": "^7.0.0",
+    "debug": "^4.3.3",
+    "err-code": "^3.0.1",
     "events": "^3.3.0",
     "hashlru": "^2.3.0",
-    "interface-datastore": "^6.0.2",
-    "it-all": "^1.0.4",
-    "it-buffer": "^0.1.2",
-    "it-drain": "^1.0.3",
-    "it-filter": "^1.0.1",
-    "it-first": "^1.0.4",
-    "it-handshake": "^2.0.0",
-    "it-length-prefixed": "^5.0.2",
-    "it-map": "^1.0.4",
-    "it-merge": "^1.0.0",
-    "it-pipe": "^1.1.0",
-    "it-take": "^1.0.0",
-    "libp2p-crypto": "^0.21.0",
-    "libp2p-interfaces": "^2.0.1",
-    "libp2p-utils": "^0.4.0",
-    "mafmt": "^10.0.0",
+    "interface-datastore": "^6.1.0",
+    "it-all": "^1.0.6",
+    "it-drain": "^1.0.5",
+    "it-filter": "^1.0.3",
+    "it-first": "^1.0.6",
+    "it-foreach": "^0.1.1",
+    "it-handshake": "^3.0.1",
+    "it-length-prefixed": "^7.0.1",
+    "it-map": "^1.0.6",
+    "it-merge": "^1.0.3",
+    "it-pipe": "^2.0.3",
+    "it-sort": "^1.0.1",
+    "it-stream-types": "^1.0.4",
+    "it-take": "^1.0.2",
+    "it-to-buffer": "^2.0.2",
     "merge-options": "^3.0.4",
-    "multiaddr": "^10.0.0",
-    "multiformats": "^9.0.0",
-    "multistream-select": "^2.0.0",
+    "mortice": "^3.0.0",
+    "multiformats": "^9.6.3",
     "mutable-proxy": "^1.0.0",
-    "nat-api": "^0.3.1",
-    "node-forge": "^0.10.0",
-    "p-any": "^3.0.0",
+    "node-forge": "^1.2.1",
     "p-fifo": "^1.0.0",
-    "p-retry": "^4.4.0",
-    "p-settle": "^4.1.1",
-    "peer-id": "^0.16.0",
-    "private-ip": "^2.1.0",
-    "protobufjs": "^6.10.2",
+    "p-retry": "^5.0.0",
+    "p-settle": "^5.0.0",
+    "private-ip": "^2.3.3",
+    "protobufjs": "^6.11.2",
     "retimer": "^3.0.0",
     "sanitize-filename": "^1.6.3",
     "set-delayed-interval": "^1.0.0",
     "streaming-iterables": "^6.0.0",
-    "timeout-abort-controller": "^2.0.0",
+    "timeout-abort-controller": "^3.0.0",
     "uint8arrays": "^3.0.0",
     "varint": "^6.0.0",
     "wherearewe": "^1.0.0",
     "xsalsa20": "^1.1.0"
   },
   "devDependencies": {
-    "@chainsafe/libp2p-noise": "^5.0.0",
+    "@achingbrain/libp2p-gossipsub": "^0.13.5",
+    "@chainsafe/libp2p-noise": "^6.0.1",
+    "@libp2p/bootstrap": "^1.0.2",
+    "@libp2p/daemon-client": "^1.0.0",
+    "@libp2p/daemon-server": "^1.0.0",
+    "@libp2p/delegated-content-routing": "^1.0.2",
+    "@libp2p/delegated-peer-routing": "^1.0.2",
+    "@libp2p/floodsub": "^1.0.2",
+    "@libp2p/interface-compliance-tests": "^1.1.20",
+    "@libp2p/interop": "^1.0.0",
+    "@libp2p/kad-dht": "^1.0.3",
+    "@libp2p/mdns": "^1.0.3",
+    "@libp2p/mplex": "^1.0.1",
+    "@libp2p/tcp": "^1.0.6",
+    "@libp2p/tracked-map": "^1.0.4",
+    "@libp2p/webrtc-star": "^1.0.3",
+    "@libp2p/websockets": "^1.0.3",
     "@nodeutils/defaults-deep": "^1.1.0",
-    "@types/es6-promisify": "^6.0.0",
-    "@types/node": "^16.0.1",
-    "@types/node-forge": "^0.10.1",
+    "@types/node": "^16.11.26",
+    "@types/node-forge": "^1.0.0",
+    "@types/p-fifo": "^1.0.0",
     "@types/varint": "^6.0.0",
-    "aegir": "^36.0.0",
+    "@types/xsalsa20": "^1.1.0",
+    "aegir": "^36.1.3",
     "buffer": "^6.0.3",
-    "datastore-core": "^6.0.7",
+    "cborg": "^1.8.1",
     "delay": "^5.0.0",
-    "into-stream": "^6.0.0",
-    "ipfs-http-client": "^54.0.2",
-    "it-concat": "^2.0.0",
-    "it-pair": "^1.0.0",
-    "it-pushable": "^1.4.0",
-    "libp2p": ".",
-    "libp2p-bootstrap": "^0.14.0",
-    "libp2p-delegated-content-routing": "^0.11.0",
-    "libp2p-delegated-peer-routing": "^0.11.1",
-    "libp2p-floodsub": "^0.28.0",
-    "libp2p-gossipsub": "^0.12.1",
-    "libp2p-interfaces-compliance-tests": "^2.0.1",
-    "libp2p-interop": "^0.5.0",
-    "libp2p-kad-dht": "^0.27.1",
-    "libp2p-mdns": "^0.18.0",
-    "libp2p-mplex": "^0.10.1",
-    "libp2p-tcp": "^0.17.0",
-    "libp2p-webrtc-star": "^0.25.0",
-    "libp2p-websockets": "^0.16.0",
+    "go-libp2p": "^0.0.6",
+    "into-stream": "^7.0.0",
+    "ipfs-http-client": "^56.0.1",
+    "it-pair": "^2.0.2",
+    "it-pushable": "^2.0.1",
     "nock": "^13.0.3",
-    "p-defer": "^3.0.0",
-    "p-times": "^3.0.0",
-    "p-wait-for": "^3.2.0",
+    "npm-run-all": "^4.1.5",
+    "p-defer": "^4.0.0",
+    "p-event": "^5.0.1",
+    "p-times": "^4.0.0",
+    "p-wait-for": "^4.1.0",
     "rimraf": "^3.0.2",
-    "sinon": "^12.0.1",
-    "util": "^0.12.3"
+    "sinon": "^13.0.1",
+    "ts-sinon": "^2.0.2"
   },
-  "contributors": [
-    "Vasco Santos <vasco.santos@moxy.studio>",
-    "David Dias <daviddias.p@gmail.com>",
-    "Jacob Heun <jacobheun@gmail.com>",
-    "Alex Potsides <alex@achingbrain.net>",
-    "Alan Shaw <alan@tableflip.io>",
-    "Cayman <caymannava@gmail.com>",
-    "Pedro Teixeira <i@pgte.me>",
-    "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
-    "Maciej Krüger <mkg20001@gmail.com>",
-    "Hugo Dias <mail@hugodias.me>",
-    "dirkmc <dirkmdev@gmail.com>",
-    "Chris Dostert <chrisdostert@users.noreply.github.com>",
-    "Volker Mische <volker.mische@gmail.com>",
-    "Robert Kiel <robert.kiel@hoprnet.org>",
-    "Richard Littauer <richard.littauer@gmail.com>",
-    "zeim839 <50573884+zeim839@users.noreply.github.com>",
-    "a1300 <matthias-knopp@gmx.net>",
-    "Ryan Bell <ryan@piing.net>",
-    "ᴠɪᴄᴛᴏʀ ʙᴊᴇʟᴋʜᴏʟᴍ <victorbjelkholm@gmail.com>",
-    "Andrew Nesbitt <andrewnez@gmail.com>",
-    "Franck Royer <franck@royer.one>",
-    "Thomas Eizinger <thomas@eizinger.io>",
-    "Giovanni T. Parra <fiatjaf@gmail.com>",
-    "acolytec3 <17355484+acolytec3@users.noreply.github.com>",
-    "Alan Smithee <ggnore.alan.smithee@gmail.com>",
-    "Elven <mon.samuel@qq.com>",
-    "Samlior <samlior@foxmail.com>",
-    "Didrik Nordström <didrik.nordstrom@gmail.com>",
-    "Soeren <nikorpoulsen@gmail.com>",
-    "Sönke Hahn <soenkehahn@gmail.com>",
-    "TJKoury <TJKoury@gmail.com>",
-    "TheStarBoys <41286328+TheStarBoys@users.noreply.github.com>",
-    "Tiago Alves <alvesjtiago@gmail.com>",
-    "XiaoZhang <zxinmyth@gmail.com>",
-    "Yusef Napora <yusef@napora.org>",
-    "Zane Starr <zcstarr@gmail.com>",
-    "ebinks <elizabethjbinks@gmail.com>",
-    "Aditya Bose <13054902+adbose@users.noreply.github.com>",
-    "isan_rivkin <isanrivkin@gmail.com>",
-    "mayerwin <mayerwin@users.noreply.github.com>",
-    "mcclure <andi.m.mcclure@gmail.com>",
-    "patrickwoodhead <91056047+patrickwoodhead@users.noreply.github.com>",
-    "phillmac <phillmac@users.noreply.github.com>",
-    "robertkiel <robert.kiel@validitylabs.org>",
-    "shresthagrawal <34920931+shresthagrawal@users.noreply.github.com>",
-    "swedneck <40505480+swedneck@users.noreply.github.com>",
-    "greenSnot <greenSnot@users.noreply.github.com>",
-    "Aleksei <vozhdb@gmail.com>",
-    "Bernd Strehl <bernd.strehl@gmail.com>",
-    "Chris Bratlien <chrisbratlien@gmail.com>",
-    "Cindy Wu <ciindy.wu@gmail.com>",
-    "Daijiro Wachi <daijiro.wachi@gmail.com>",
-    "Diogo Silva <fsdiogo@gmail.com>",
-    "Dmitriy Ryajov <dryajov@gmail.com>",
-    "Ethan Lam <elmemphis2000@gmail.com>",
-    "Fei Liu <liu.feiwood@gmail.com>",
-    "Felipe Martins <felipebrasil93@gmail.com>",
-    "Florian-Merle <florian.david.merle@gmail.com>",
-    "Francis Gulotta <wizard@roborooter.com>",
-    "Guy Sviry <32539816+guysv@users.noreply.github.com>",
-    "Henrique Dias <hacdias@gmail.com>",
-    "Irakli Gozalishvili <rfobic@gmail.com>",
-    "Joel Gustafson <joelg@mit.edu>",
-    "John Rees <johnrees@users.noreply.github.com>",
-    "João Santos <joaosantos15@users.noreply.github.com>",
-    "Julien Bouquillon <contact@revolunet.com>",
-    "Kevin Kwok <antimatter15@gmail.com>",
-    "Kevin Lacker <lacker@gmail.com>",
-    "Lars Gierth <lgierth@users.noreply.github.com>",
-    "Leask Wong <i@leaskh.com>",
-    "Marcin Tojek <mtojek@users.noreply.github.com>",
-    "Marston Connell <34043723+TheMarstonConnell@users.noreply.github.com>",
-    "Michael Burns <5170+mburns@users.noreply.github.com>",
-    "Miguel Mota <miguelmota2@gmail.com>",
-    "Nuno Nogueira <nunofmn@gmail.com>",
-    "Philipp Muens <raute1337@gmx.de>",
-    "RasmusErik Voel Jensen <github@solsort.com>",
-    "Smite Chow <xiaopengyou@live.com>"
-  ]
+  "browser": {
+    "nat-api": false
+  }
 }
