@@ -1,26 +1,23 @@
 // eslint-disable-next-line
 'use strict'
 
-const Libp2p = require('libp2p')
-const Websockets = require('libp2p-websockets')
-const WebSocketStar = require('libp2p-websocket-star')
-const WebRTCStar = require('libp2p-webrtc-star')
-const MPLEX = require('libp2p-mplex')
-const SECIO = require('libp2p-secio')
-const KadDHT = require('libp2p-kad-dht')
-const DelegatedPeerRouter = require('libp2p-delegated-peer-routing')
-const DelegatedContentRouter = require('libp2p-delegated-content-routing')
+import { createLibp2p } from 'libp2p'
+import { WebSockets } from '@libp2p/websockets'
+import { WebRTCStar } from '@libp2p/webrtc-star'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { DelegatedPeerRouting } from '@libp2p/delegated-peer-routing'
+import { DelegatedContentRouting } from '@libp2p/delegated-content-routing'
 
 export default function Libp2pBundle ({peerInfo, peerBook}) {
-  const wrtcstar = new WebRTCStar({id: peerInfo.id})
-  const wsstar = new WebSocketStar({id: peerInfo.id})
+  const wrtcstar = new WebRTCStar()
   const delegatedApiOptions = {
     host: '0.0.0.0',
     protocol: 'http',
     port: '8080'
   }
 
-  return new Libp2p({
+  return createLibp2p({
     peerInfo,
     peerBook,
     // Lets limit the connection managers peers and have it check peer health less frequently
@@ -28,48 +25,29 @@ export default function Libp2pBundle ({peerInfo, peerBook}) {
       maxPeers: 10,
       pollInterval: 5000
     },
-    modules: {
-      contentRouting: [
-        new DelegatedContentRouter(peerInfo.id, delegatedApiOptions)
-      ],
-      peerRouting: [
-        new DelegatedPeerRouter(delegatedApiOptions)
-      ],
-      peerDiscovery: [
-        wrtcstar.discovery,
-        wsstar.discovery
-      ],
-      transport: [
-        wrtcstar,
-        wsstar,
-        Websockets
-      ],
-      streamMuxer: [
-        MPLEX
-      ],
-      connEncryption: [
-        SECIO
-      ],
-      dht: KadDHT
+    contentRouting: [
+      new DelegatedPeerRouting(peerInfo.id, delegatedApiOptions)
+    ],
+    peerRouting: [
+      new DelegatedContentRouting(delegatedApiOptions)
+    ],
+    transports: [
+      wrtcstar,
+      new WebSockets()
+    ],
+    streamMuxers: [
+      new Mplex()
+    ],
+    connectionEncryption: [
+      new Noise()
+    ],
+    connectionManager: {
+      autoDial: false
     },
-    config: {
-      peerDiscovery: {
-        autoDial: false,
-        webrtcStar: {
-          enabled: false
-        },
-        websocketStar: {
-          enabled: false
-        }
-      },
-      dht: {
+    relay: {
+      enabled: true,
+      hop: {
         enabled: false
-      },
-      relay: {
-        enabled: true,
-        hop: {
-          enabled: false
-        }
       }
     }
   })
