@@ -1,23 +1,21 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const MPLEX = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-
-const pipe = require('it-pipe')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { pipe } from 'it-pipe'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [MPLEX],
-      connEncryption: [NOISE]
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()]
   })
 
   await node.start()
@@ -32,14 +30,14 @@ const createNode = async () => {
   ])
 
   // Add node's 2 data to the PeerStore
-  node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
-  
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
+
   node1.handle('/node-1', ({ stream }) => {
     pipe(
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(msg.toString())
+          console.log(uint8ArrayToString(msg))
         }
       }
     )
@@ -50,7 +48,7 @@ const createNode = async () => {
       stream,
       async function (source) {
         for await (const msg of source) {
-          console.log(msg.toString())
+          console.log(uint8ArrayToString(msg))
         }
       }
     )
@@ -58,13 +56,13 @@ const createNode = async () => {
 
   const { stream: stream1 } = await node1.dialProtocol(node2.peerId, ['/node-2'])
   await pipe(
-    ['from 1 to 2'],
+    [uint8ArrayFromString('from 1 to 2')],
     stream1
   )
 
   const { stream: stream2 } = await node2.dialProtocol(node1.peerId, ['/node-1'])
   await pipe(
-    ['from 2 to 1'],
+    [uint8ArrayFromString('from 2 to 1')],
     stream2
   )
 })();
