@@ -1,25 +1,23 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../../')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const Gossipsub = require('libp2p-gossipsub')
-const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
-const { toString: uint8ArrayToString } = require('uint8arrays/to-string')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { Gossipsub } from '@achingbrain/libp2p-gossipsub'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
+import { CustomEvent } from '@libp2p/interfaces'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      pubsub: Gossipsub
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncryption: [new Noise()],
+    pubsub: new Gossipsub()
   })
 
   await node.start()
@@ -36,28 +34,26 @@ const createNode = async () => {
   ])
 
   // node1 conect to node2 and node2 conect to node3
-  node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
   await node1.dial(node2.peerId)
 
-  node2.peerStore.addressBook.set(node3.peerId, node3.multiaddrs)
+  await node2.peerStore.addressBook.set(node3.peerId, node3.getMultiaddrs())
   await node2.dial(node3.peerId)
 
   //subscribe
-  node1.pubsub.on(topic, (msg) => {
+  node1.pubsub.addEventListener(topic, (evt) => {
     // Will not receive own published messages by default
-    console.log(`node1 received: ${uint8ArrayToString(msg.data)}`)
+    console.log(`node1 received: ${uint8ArrayToString(evt.detail.data)}`)
   })
   await node1.pubsub.subscribe(topic)
 
-  node2.pubsub.on(topic, (msg) => {
-    console.log(`node2 received: ${uint8ArrayToString(msg.data)}`)
+  node2.pubsub.addEventListener(topic, (evt) => {
+    console.log(`node2 received: ${uint8ArrayToString(evt.detail.data)}`)
   })
-  await node2.pubsub.subscribe(topic)
 
-  node3.pubsub.on(topic, (msg) => {
-    console.log(`node3 received: ${uint8ArrayToString(msg.data)}`)
+  node3.pubsub.addEventListener(topic, (evt) => {
+    console.log(`node3 received: ${uint8ArrayToString(evt.detail.data)}`)
   })
-  await node3.pubsub.subscribe(topic)
 
   const validateFruit = (msgTopic, msg) => {
     const fruit = uint8ArrayToString(msg.data)
@@ -79,7 +75,7 @@ const createNode = async () => {
   // car is not a fruit !
   setInterval(() => {
     console.log('############## fruit ' + myFruits[count] + ' ##############')
-    node1.pubsub.publish(topic, uint8ArrayFromString(myFruits[count]))
+    node1.pubsub.dispatchEvent(new CustomEvent<Uint8Array>(topic, { detail: uint8ArrayFromString(myFruits[count]) }))
     count++
     if (count == myFruits.length) {
       count = 0
