@@ -1,28 +1,26 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../..')
-const TCP = require('libp2p-tcp')
-const WebSockets = require('libp2p-websockets')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const MPLEX = require('libp2p-mplex')
-
-const pipe = require('it-pipe')
+import { createLibp2p } from 'libp2p'
+import { TCP } from '@libp2p/tcp'
+import { WebSockets } from '@libp2p/websockets'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { Mplex } from '@libp2p/mplex'
+import { pipe } from 'it-pipe'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 
 const createNode = async (transports, addresses = []) => {
   if (!Array.isArray(addresses)) {
     addresses = [addresses]
   }
 
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: addresses
     },
-    modules: {
-      transport: transports,
-      connEncryption: [NOISE],
-      streamMuxer: [MPLEX]
-    }
+    transports: transports,
+    connectionEncryption: [new Noise()],
+    streamMuxers: [new Mplex()]
   })
 
   await node.start()
@@ -31,7 +29,7 @@ const createNode = async (transports, addresses = []) => {
 
 function printAddrs(node, number) {
   console.log('node %s is listening on:', number)
-  node.multiaddrs.forEach((ma) => console.log(`${ma.toString()}/p2p/${node.peerId.toB58String()}`))
+  node.getMultiaddrs().forEach((ma) => console.log(ma.toString()))
 }
 
 function print ({ stream }) {
@@ -39,7 +37,7 @@ function print ({ stream }) {
     stream,
     async function (source) {
       for await (const msg of source) {
-        console.log(msg.toString())
+        console.log(uint8ArrayToString(msg))
       }
     }
   )
@@ -47,9 +45,9 @@ function print ({ stream }) {
 
 ;(async () => {
   const [node1, node2, node3] = await Promise.all([
-    createNode([TCP], '/ip4/0.0.0.0/tcp/0'),
-    createNode([TCP, WebSockets], ['/ip4/0.0.0.0/tcp/0', '/ip4/127.0.0.1/tcp/10000/ws']),
-    createNode([WebSockets], '/ip4/127.0.0.1/tcp/20000/ws')
+    createNode([new TCP()], '/ip4/0.0.0.0/tcp/0'),
+    createNode([new TCP(), new WebSockets()], ['/ip4/0.0.0.0/tcp/0', '/ip4/127.0.0.1/tcp/10000/ws']),
+    createNode([new WebSockets()], '/ip4/127.0.0.1/tcp/20000/ws')
   ])
 
   printAddrs(node1, '1')
@@ -60,28 +58,28 @@ function print ({ stream }) {
   node2.handle('/print', print)
   node3.handle('/print', print)
 
-  node1.peerStore.addressBook.set(node2.peerId, node2.multiaddrs)
-  node2.peerStore.addressBook.set(node3.peerId, node3.multiaddrs)
-  node3.peerStore.addressBook.set(node1.peerId, node1.multiaddrs)
+  await node1.peerStore.addressBook.set(node2.peerId, node2.getMultiaddrs())
+  await node2.peerStore.addressBook.set(node3.peerId, node3.getMultiaddrs())
+  await node3.peerStore.addressBook.set(node1.peerId, node1.getMultiaddrs())
 
   // node 1 (TCP) dials to node 2 (TCP+WebSockets)
   const { stream } = await node1.dialProtocol(node2.peerId, '/print')
   await pipe(
-    ['node 1 dialed to node 2 successfully'],
+    [uint8ArrayFromString('node 1 dialed to node 2 successfully')],
     stream
   )
 
   // node 2 (TCP+WebSockets) dials to node 2 (WebSockets)
   const { stream: stream2 } = await node2.dialProtocol(node3.peerId, '/print')
   await pipe(
-    ['node 2 dialed to node 3 successfully'],
+    [uint8ArrayFromString('node 2 dialed to node 3 successfully')],
     stream2
   )
 
   // node 3 (listening WebSockets) can dial node 1 (TCP)
   try {
     await node3.dialProtocol(node1.peerId, '/print')
-  } catch (/** @type {any} */ err) {
+  } catch (err) {
     console.log('node 3 failed to dial to node 1 with:', err.message)
   }
 })();
