@@ -1,15 +1,13 @@
-'use strict'
-
-const execa = require('execa')
-const fs = require('fs-extra')
-const which = require('which')
+import execa from 'execa'
+import fs from 'fs-extra'
+import which from 'which'
 
 async function isExecutable (command) {
   try {
     await fs.access(command, fs.constants.X_OK)
 
     return true
-  } catch (/** @type {any} */ err) {
+  } catch (err) {
     if (err.code === 'ENOENT') {
       return isExecutable(await which(command))
     }
@@ -22,7 +20,7 @@ async function isExecutable (command) {
   }
 }
 
-async function waitForOutput (expectedOutput, command, args = [], opts = {}) {
+export async function waitForOutput (expectedOutput, command, args = [], opts = {}) {
   if (!await isExecutable(command)) {
     args.unshift(command)
     command = 'node'
@@ -30,7 +28,7 @@ async function waitForOutput (expectedOutput, command, args = [], opts = {}) {
 
   const proc = execa(command, args, opts)
   let output = ''
-  let time = 120000
+  let time = 600000
 
   let timeout = setTimeout(() => {
     throw new Error(`Did not see "${expectedOutput}" in output from "${[command].concat(args).join(' ')}" after ${time/1000}s`)
@@ -49,13 +47,9 @@ async function waitForOutput (expectedOutput, command, args = [], opts = {}) {
 
   try {
     await proc
-  } catch (/** @type {any} */ err) {
+  } catch (err) {
     if (!err.killed) {
       throw err
     }
   }
-}
-
-module.exports = {
-  waitForOutput
 }
